Polish Perfetto iFrame lifecycle
