+++ packages/devtools_app/lib/src/screens/performance/panes/timeline_events/perfetto/_perfetto_controller_web.dart
@@ -3,13 +3,11 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:convert';
 import 'dart:html' as html;
 import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart';
 
-import '../../../../../primitives/auto_dispose.dart';
 import '../../../../../primitives/trace_event.dart';
 import '../../../../../primitives/utils.dart';
 import '../../../../../shared/globals.dart';
@@ -36,14 +34,12 @@ const _debugUseLocalPerfetto = false;
 /// [_viewIdIncrementer] is used to create.
 var _viewIdIncrementer = 0;
 
-class PerfettoControllerImpl extends PerfettoController
-    with AutoDisposeControllerMixin {
+class PerfettoControllerImpl extends PerfettoController {
   PerfettoControllerImpl(
     super.performanceController,
     super.timelineEventsController,
   );
 
-  @override
   late final viewId = 'embedded-perfetto-${_viewIdIncrementer++}';
 
   /// Url when running Perfetto locally following the instructions here:
@@ -52,41 +48,7 @@ class PerfettoControllerImpl extends PerfettoController
 
   static const _embeddedModeQuery = '?mode=embedded&hideSidebar=true';
 
-  static const _perfettoPing = 'PING';
-
-  static const _perfettoPong = 'PONG';
-
-  /// Id for a [postMessage] request that is sent before trying to change the
-  /// DevTools theme (see [_devtoolsThemeChange]).
-  ///
-  /// Once the DevTools theme handler in the bundled Perfetto web app has been
-  /// registered, a "pong" event [_devtoolsThemePong] will be returned, at which
-  /// point we can safely change the theme [_devtoolsThemeChange].
-  ///
-  /// This message must be sent with the argument 'perfettoIgnore' set to true
-  /// so that the message handler in the Perfetto codebase
-  /// [post_message_handler.ts] will not try to handle this message and warn
-  /// "Unknown postMessage() event received".
-  static const _devtoolsThemePing = 'DART-DEVTOOLS-THEME-PING';
-
-  /// Id for a [postMessage] response that should be received when the DevTools
-  /// theme handler has been registered.
-  ///
-  /// We will send a "ping" event [_devtoolsThemePing] to the DevTools theme
-  /// handler in the bundled Perfetto web app, and the handler will return this
-  /// "pong" event when it is ready. We must wait for this event to be returned
-  /// before we can send a theme change request [_devtoolsThemeChange].
-  static const _devtoolsThemePong = 'DART-DEVTOOLS-THEME-PONG';
-
-  /// Id for a [postMessage] request that is sent on DevTools theme changes.
-  ///
-  /// This message must be sent with the argument 'perfettoIgnore' set to true
-  /// so that the message handler in the Perfetto codebase
-  /// [post_message_handler.ts] will not try to handle this message and warn
-  /// "Unknown postMessage() event received".
-  static const _devtoolsThemeChange = 'DART-DEVTOOLS-THEME-CHANGE';
-
-  String get _perfettoUrl {
+  String get perfettoUrl {
     if (_debugUseLocalPerfetto) {
       return _debugPerfettoUrl;
     }
@@ -96,31 +58,17 @@ class PerfettoControllerImpl extends PerfettoController
     return '$baseUrl$_embeddedModeQuery';
   }
 
+  html.IFrameElement get perfettoIFrame => _perfettoIFrame;
+
   late final html.IFrameElement _perfettoIFrame;
 
-  /// Completes when the perfetto iFrame has recevied the first event on the
-  /// 'onLoad' stream.
-  late final Completer<void> _perfettoIFrameReady;
+  ValueListenable<List<TraceEventWrapper>> get activeTraceEvents =>
+      _activeTraceEvents;
+  final _activeTraceEvents = ValueNotifier<List<TraceEventWrapper>>([]);
 
-  /// Completes when the Perfetto postMessage handler is ready, which is
-  /// signaled by receiving a [_perfettoPong] event in response to sending a
-  /// [_perfettoPing] event.
-  late final Completer<void> _perfettoHandlerReady;
-
-  /// Completes when the DevTools theme postMessage handler is ready, which is
-  /// signaled by receiving a [_devtoolsThemePong] event in response to sending
-  /// a [_devtoolsThemePing] event.
-  late final Completer<void> _devtoolsThemeHandlerReady;
-
-  /// Timer that will poll until [_perfettoHandlerReady] is complete or until
-  /// [_pollUntilReadyTimeout] has passed.
-  Timer? _pollForPerfettoHandlerReady;
-
-  /// Timer that will poll until [_devtoolsThemeHandlerReady] is complete or
-  /// until [_pollUntilReadyTimeout] has passed.
-  Timer? _pollForThemeHandlerReady;
-
-  static const _pollUntilReadyTimeout = Duration(seconds: 10);
+  ValueListenable<TimeRange?> get activeScrollToTimeRange =>
+      _activeScrollToTimeRange;
+  final _activeScrollToTimeRange = ValueNotifier<TimeRange?>(null);
 
   /// Trace events that we should load, but have not yet since the trace viewer
   /// is not visible (i.e. [TimelineEventsController.isActiveFeature] is false).
@@ -130,32 +78,26 @@ class PerfettoControllerImpl extends PerfettoController
   /// is not visible (i.e. [TimelineEventsController.isActiveFeature] is false).
   TimeRange? pendingScrollToTimeRange;
 
-  /// Boolean value representing the pending theme change we that we should
-  /// apply, but have not yet since the trace viewer is not visible (i.e.
-  /// [TimelineEventsController.isActiveFeature] is false).
-  bool? pendingLoadDarkMode;
+  bool _initialized = false;
 
   @override
   void init() {
-    _perfettoIFrameReady = Completer();
-    _perfettoHandlerReady = Completer();
-    _devtoolsThemeHandlerReady = Completer();
+    assert(
+      !_initialized,
+      'PerfettoController.init() should only be called once.',
+    );
+    _initialized = true;
+
     _perfettoIFrame = html.IFrameElement()
       // This url is safe because we built it ourselves and it does not include
       // any user input.
       // ignore: unsafe_html
-      ..src = _perfettoUrl
+      ..src = perfettoUrl
       ..allow = 'usb';
     _perfettoIFrame.style
       ..border = 'none'
       ..height = '100%'
       ..width = '100%';
-
-    unawaited(
-      _perfettoIFrame.onLoad.first.then((_) {
-        _perfettoIFrameReady.complete();
-      }),
-    );
 
     // ignore: undefined_prefixed_name
     final registered = ui.platformViewRegistry.registerViewFactory(
@@ -163,22 +105,11 @@ class PerfettoControllerImpl extends PerfettoController
       (int viewId) => _perfettoIFrame,
     );
     assert(registered, 'Failed to register view factory for $viewId.');
-
-    html.window.addEventListener('message', _handleMessage);
-
-    if (isExternalBuild) {
-      unawaited(_loadStyle(preferences.darkModeTheme.value));
-      addAutoDisposeListener(preferences.darkModeTheme, () async {
-        await _loadStyle(preferences.darkModeTheme.value);
-      });
-    }
   }
 
   @override
   Future<void> onBecomingActive() async {
-    if (pendingLoadDarkMode != null) {
-      unawaited(_loadStyle(pendingLoadDarkMode!));
-    }
+    assert(timelineEventsController.isActiveFeature);
     if (pendingTraceEventsToLoad != null) {
       await loadTrace(pendingTraceEventsToLoad!);
       pendingTraceEventsToLoad = null;
@@ -196,23 +127,7 @@ class PerfettoControllerImpl extends PerfettoController
       return;
     }
     pendingTraceEventsToLoad = null;
-
-    await _pingPerfettoUntilReady();
-
-    final encodedJson = jsonEncode({
-      'traceEvents': devToolsTraceEvents
-          .map((eventWrapper) => eventWrapper.event.json)
-          .toList(),
-    });
-    final buffer = Uint8List.fromList(encodedJson.codeUnits);
-
-    _postMessage({
-      'perfetto': {
-        'buffer': buffer,
-        'title': 'DevTools timeline trace',
-        'keepApiOpen': true,
-      }
-    });
+    _activeTraceEvents.value = List.from(devToolsTraceEvents);
   }
 
   @override
@@ -222,131 +137,11 @@ class PerfettoControllerImpl extends PerfettoController
       return;
     }
     pendingScrollToTimeRange = null;
-
-    if (!timeRange.isWellFormed) {
-      notificationService.push(
-        'No timeline events available for the selected frame. Timeline '
-        'events occurred too long ago before DevTools could access them. '
-        'To avoid this, open the DevTools Performance page sooner.',
-      );
-      return;
-    }
-    await _pingPerfettoUntilReady();
-    _postMessage({
-      'perfetto': {
-        // Pass the values to Perfetto in seconds.
-        'timeStart': timeRange.start!.inMicroseconds / 1000000,
-        'timeEnd': timeRange.end!.inMicroseconds / 1000000,
-        // The time range should take up 80% of the visible window.
-        'viewPercentage': 0.8,
-      }
-    });
-  }
-
-  Future<void> _loadStyle(bool darkMode) async {
-    if (!isExternalBuild) return;
-    if (!timelineEventsController.isActiveFeature) {
-      pendingLoadDarkMode = darkMode;
-      return;
-    }
-    pendingLoadDarkMode = null;
-
-    // This message will be handled by [devtools_theme_handler.js], which is
-    // included in the Perfetto build inside [packages/perfetto_compiled/dist].
-    await _pingDevToolsThemeHandlerUntilReady();
-    _postMessageWithId(
-      _devtoolsThemeChange,
-      perfettoIgnore: true,
-      args: {
-        'theme': '${darkMode ? 'dark' : 'light'}',
-      },
-    );
-  }
-
-  void _postMessage(dynamic message) async {
-    await _perfettoIFrameReady.future;
-    assert(
-      _perfettoIFrame.contentWindow != null,
-      'Something went wrong. The iFrame\'s contentWindow is null after the'
-      ' _perfettoIFrameReady future completed.',
-    );
-    _perfettoIFrame.contentWindow!.postMessage(
-      message,
-      _perfettoUrl,
-    );
-  }
-
-  void _postMessageWithId(
-    String id, {
-    Map<String, dynamic> args = const {},
-    bool perfettoIgnore = false,
-  }) {
-    final message = <String, dynamic>{
-      'msgId': id,
-      if (perfettoIgnore) 'perfettoIgnore': true,
-    }..addAll(args);
-    _postMessage(message);
-  }
-
-  void _handleMessage(html.Event e) {
-    if (e is html.MessageEvent) {
-      if (e.data == _perfettoPong && !_perfettoHandlerReady.isCompleted) {
-        _perfettoHandlerReady.complete();
-      }
-
-      if (e.data == _devtoolsThemePong &&
-          !_devtoolsThemeHandlerReady.isCompleted) {
-        _devtoolsThemeHandlerReady.complete();
-      }
-    }
-  }
-
-  Future<void> _pingPerfettoUntilReady() async {
-    if (!_perfettoHandlerReady.isCompleted) {
-      _pollForPerfettoHandlerReady =
-          Timer.periodic(const Duration(milliseconds: 200), (_) async {
-        // Once the Perfetto UI is ready, Perfetto will receive this 'PING'
-        // message and return a 'PONG' message, handled in [_handleMessage].
-        _postMessage(_perfettoPing);
-      });
-
-      await _perfettoHandlerReady.future.timeout(
-        _pollUntilReadyTimeout,
-        onTimeout: () => _pollForPerfettoHandlerReady?.cancel(),
-      );
-      _pollForPerfettoHandlerReady?.cancel();
-    }
-  }
-
-  Future<void> _pingDevToolsThemeHandlerUntilReady() async {
-    if (!isExternalBuild) return;
-    if (!_devtoolsThemeHandlerReady.isCompleted) {
-      _pollForThemeHandlerReady =
-          Timer.periodic(const Duration(milliseconds: 200), (_) async {
-        // Once [devtools_theme_handler.js] is ready, it will receive this
-        // 'PING-DEVTOOLS-THEME' message and return a 'PONG-DEVTOOLS-THEME'
-        // message, handled in [_handleMessage].
-        _postMessageWithId(_devtoolsThemePing, perfettoIgnore: true);
-      });
-
-      await _devtoolsThemeHandlerReady.future.timeout(
-        _pollUntilReadyTimeout,
-        onTimeout: () => _pollForThemeHandlerReady?.cancel(),
-      );
-      _pollForThemeHandlerReady?.cancel();
-    }
+    _activeScrollToTimeRange.value = timeRange;
   }
 
   @override
   Future<void> clear() async {
     await loadTrace([]);
-  }
-
-  @override
-  void dispose() {
-    html.window.removeEventListener('message', _handleMessage);
-    _pollForPerfettoHandlerReady?.cancel();
-    _pollForThemeHandlerReady?.cancel();
-    super.dispose();
   }
 }
